Changed deserialization and refactorings
