+++ src/Components/Endpoints/src/Microsoft.AspNetCore.Components.Endpoints.csproj
@@ -40,6 +40,7 @@
     <Compile Include="$(ComponentsSharedSourceRoot)src\Reflection\PropertyGetter.cs" />
     <Compile Include="$(ComponentsSharedSourceRoot)src\Reflection\PropertySetter.cs" />
     <Compile Include="$(ComponentsSharedSourceRoot)src\RenderFragmentSerializer.cs" />
+    <Compile Include="$(ComponentsSharedSourceRoot)src\RenderFragmentCapture.cs" />
 
     <Compile Include="$(SharedSourceRoot)PropertyHelper\**\*.cs" />
