+++ src/Components/Server/src/Circuits/ComponentParameterDeserializer.cs
@@ -49,36 +49,41 @@ internal sealed partial class ComponentParameterDeserializer
                 Log.IncompleteParameterDefinition(_logger, definition.Name, definition.TypeName, definition.Assembly);
                 return false;
             }
+            else if (definition.TypeName == typeof(SerializedRenderFragment).FullName
+                && definition.Assembly == "Microsoft.AspNetCore.Components.Endpoints")
+            {
+                try
+                {
+                    var value = (JsonElement)parameterValues[i];
+                    var serialized = JsonSerializer.Deserialize<SerializedRenderFragment>(
+                        value.GetRawText(),
+                        ServerComponentSerializationSettings.JsonSerializationOptions);
+                    parametersDictionary.Add(definition.Name, RenderFragmentSerializer.Deserialize(serialized!.Nodes));
+                }
+                catch (Exception e)
+                {
+                    Log.InvalidParameterValue(_logger, definition.Name, definition.TypeName, definition.Assembly, e);
+                    return false;
+                }
+            }
             else
             {
+                var parameterType = _parametersCache.GetParameterType(definition.Assembly, definition.TypeName);
+                if (parameterType == null)
+                {
+                    Log.InvalidParameterType(_logger, definition.Name, definition.TypeName, definition.Assembly);
+                    return false;
+                }
                 try
                 {
                     // At this point we know the parameter is not null, as we don't serialize the type name or the assembly name
                     // for null parameters.
                     var value = (JsonElement)parameterValues[i];
-
-                    if (definition.TypeName == typeof(SerializedRenderFragment).FullName)
-                    {
-                        var serialized = JsonSerializer.Deserialize<SerializedRenderFragment>(
-                            value.GetRawText(),
-                            ServerComponentSerializationSettings.JsonSerializationOptions);
-                        parametersDictionary.Add(definition.Name, RenderFragmentSerializer.Deserialize(serialized!.Nodes));
-                    }
-                    else
-                    {
-                        var parameterType = _parametersCache.GetParameterType(definition.Assembly, definition.TypeName);
-                        if (parameterType is null)
-                        {
-                            Log.InvalidParameterType(_logger, definition.Name, definition.TypeName, definition.Assembly);
-                            return false;
-                        }
-
-                        var parameterValue = JsonSerializer.Deserialize(
-                            value.GetRawText(),
-                            parameterType,
-                            ServerComponentSerializationSettings.JsonSerializationOptions);
-                        parametersDictionary.Add(definition.Name, parameterValue);
-                    }
+                    var parameterValue = JsonSerializer.Deserialize(
+                        value.GetRawText(),
+                        parameterType,
+                        ServerComponentSerializationSettings.JsonSerializationOptions);
+                    parametersDictionary.Add(definition.Name, parameterValue);
                 }
                 catch (Exception e)
                 {
